Simplify de-conflicting JSON schema name
This commit simplifies what is and can be de-conflicted. We don't
actually generate top-level simple types, so a bunch of code for that
was removed. Rather than auto-incrementing a number to de-conflict, we
will now append a formatted version of the shape ID's namespace to the
conflicting shape name to make the results more deterministic and
less-likely to conflict with explicitly named shapes (like List2?).